[SPR-65] feat: 통계기록반환
- 쿼리에서 바로 dto를 만들어 반환하려던 방법을 버리고 Tuple로 받은 후 각 값에 맞게 mapping해준 후 dto로 변환하여 반환하였습니다.
- 쿼리 후 totalTime은 Double로 반환되기 때문에 LocalTime으로 convert해주는 로직을 작성하였습니다.

+++ src/main/java/com/climeet/climeet_backend/domain/climbingrecord/dto/ClimbingRecordResponseDto.java
@@ -42,15 +42,9 @@ public class ClimbingRecordResponseDto {
     public static class ClimbingRecordStatisticsInfo {
 
         private LocalTime time;
-        private Integer totalCompletedCount;
-        private Integer attemptRouteCount;
-        private Integer avgDifficulty;
+        private Long totalCompletedCount;
+        private Long attemptRouteCount;
+        private Long avgDifficulty;
 
-//        public ClimbingRecordStatisticsInfo(Long time, Integer totalCompletedCount, Integer attemptRouteCount, Integer avgDifficulty) {
-//            this.time = time;
-//            this.totalCompletedCount = totalCompletedCount;
-//            this.attemptRouteCount = attemptRouteCount;
-//            this.avgDifficulty = avgDifficulty;
-//        }
     }
 }
